# be/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  docs.Meta:
    properties:
      request_id:
        example: req_abc123
        type: string
      timestamp:
        example: "2025-12-08T10:00:00Z"
        type: string
    type: object
  docs.SuccessResponse:
    description: Standard success response format
    properties:
      data: {}
      meta:
        $ref: '#/definitions/docs.Meta'
      success:
        example: true
        type: boolean
    type: object
...
  errors.ErrorBody:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/errors.ErrorDetail'
        type: array
      message:
        type: string
    type: object
  errors.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorBody'
      meta:
        $ref: '#/definitions/errors.ResponseMeta'
      success:
        type: boolean
    type: object
  errors.ResponseMeta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
  map_string_array_string:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
  response.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  response.PaginatedResponse-any:
    properties:
      data:
        items: {}
        type: array
      meta:
        $ref: '#/definitions/response.Meta'
      pagination:
        $ref: '#/definitions/response.Pagination'
      success:
        type: boolean
    type: object
...
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response-any:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
  response.Response-map_string_array_string:
    properties:
      data:
        $ref: '#/definitions/map_string_array_string'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
  response.Response-map_string_string:
    properties:
      data:
        $ref: '#/definitions/map_string_string'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8001
info:
  contact:
    email: support@ngasihtau.com
    name: NgasihTau API Support
  description: |-
    NgasihTau is a learning platform API that enables teachers to create Knowledge Pods for sharing learning materials with students.
    The platform uses a microservices architecture with Go as the primary backend language.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NgasihTau API
  version: "1.0"
paths:
...
  /search:
    get:
      consumes:
      - application/json
      description: Search for pods and materials using full-text search
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by type (pod, material)
        in: query
        name: type
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by file type (pdf, docx, pptx)
        in: query
        name: file_type
        type: string
      - description: Filter by pod ID
        format: uuid
        in: query
        name: pod_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/response.PaginatedResponse-any'
      security:
      - BearerAuth: []
      summary: Full-text search
      tags:
      - Search
  /search/history:
    delete:
      consumes:
      - application/json
      description: Clear the authenticated user's search history
      produces:
      - application/json
      responses:
        "200":
          description: History cleared
          schema:
            $ref: '#/definitions/response.Response-map_string_string'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Clear search history
      tags:
      - Search
    get:
      consumes:
      - application/json
      description: Get the authenticated user's search history
      parameters:
      - default: 20
        description: Maximum results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search history
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get search history
      tags:
      - Search
  /search/hybrid:
    get:
      consumes:
      - application/json
      description: Combine keyword and semantic search for best results
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by file type
        in: query
        name: file_type
        type: string
      - description: Filter by pod ID
        format: uuid
        in: query
        name: pod_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - default: 0.3
        description: Weight for semantic results (0-1)
        in: query
        name: semantic_weight
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Hybrid search results
          schema:
            $ref: '#/definitions/response.PaginatedResponse-any'
      security:
      - BearerAuth: []
      summary: Hybrid search
      tags:
      - Search
  /search/popular:
    get:
      consumes:
      - application/json
      description: Get materials ranked by all-time engagement
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - default: 20
        description: Maximum results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Popular materials
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get popular materials
      tags:
      - Search
  /search/semantic:
    get:
      consumes:
      - application/json
      description: Search using natural language with vector similarity
      parameters:
      - description: Natural language query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by pod ID
        format: uuid
        in: query
        name: pod_id
        type: string
      - default: 10
        description: Maximum results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Minimum similarity score
        in: query
        name: min_score
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Semantic search results
          schema:
            $ref: '#/definitions/response.Response-any'
        "400":
          description: Query parameter required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Semantic search
      tags:
      - Search
  /search/suggestions:
    get:
      consumes:
      - application/json
      description: Get autocomplete suggestions based on query prefix
      parameters:
      - description: Query prefix
        in: query
        name: q
        required: true
        type: string
      - default: 10
        description: Maximum suggestions
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Suggestions
          schema:
            $ref: '#/definitions/response.Response-map_string_array_string'
        "400":
          description: Query parameter required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get search suggestions
      tags:
      - Search
  /search/trending:
    get:
      consumes:
      - application/json
      description: Get materials ranked by recent engagement (last 7 days)
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - default: 20
        description: Maximum results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trending materials
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get trending materials
      tags:
      - Search
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"